==> 03a_select_regions_to_do_TOO.ipynb <==
import pandas as pd 
import numpy as np 
import seaborn as sns 
import matplotlib.pyplot as plt 
from tqdm import tqdm 
import warnings
warnings.filterwarnings("ignore")
import os
import pathlib
import random 
import argparse
from scipy import optimize

inputdir = "/Volumes/HNWD02/outdir/output_methyl_ReadBased_Models/01_output/TSMA_panel"
path_to_all_fa = "/Volumes/HNSD01/storage/ref/hg19"
outdir = "/Volumes/HNSD01/outdir"
outputdir = os.path.join(outdir, "output_methyl_ReadBased_Models")
path_to_01_output = os.path.join(outputdir, "01_output")
path_to_02_output = os.path.join(outputdir, "02_output")

metadata = pd.read_excel("/Users/hieunguyen/src/tsma_micro_atlas/metadata/metadata_TSMA_R8281_R8282.xlsx")
paneldf = pd.read_excel("./assets/12967_2024_5416_MOESM1_ESM.xlsx")
all_regions = paneldf.Region_name.unique()

path_to_03_output = os.path.join(outputdir, "03_output")
path_to_03a_output = os.path.join(outputdir, "03_output")
os.system(f"mkdir -p {path_to_03a_output}")

all_predictiondf = pd.DataFrame(
    data = ['Breast', 'CRC', 'Gastric', 'Liver', 'Lung', 'WBC', 'no data'],
    columns = ['prediction']
)



if os.path.isfile(os.path.join(path_to_03a_output, f"LOOCV_res.xlsx")) == False:
    for sampleid in metadata.LABCODE.unique():
        print(f"working on sample {sampleid}...")
        all_preds = [
                item for item in pathlib.Path(path_to_03_output).glob(f"*{sampleid}/microAtlas_data/region_*/*_methyl_pattern_by_tissue_type.csv")
            ]

        cosdf = pd.DataFrame()
        for file in tqdm(all_preds):
            region_name = file.name.replace("_methyl_pattern_by_tissue_type.csv", "")
            tmpdf = pd.read_csv(file, index_col=0)

            from sklearn.metrics.pairwise import cosine_similarity

            cosine_sim = cosine_similarity(tmpdf.T)
            cosine_sim_df = pd.DataFrame(cosine_sim, 
                                        index=tmpdf.columns, 
                                        columns=tmpdf.columns)
            sum_cos_sim = np.sum(cosine_sim_df.abs().sum().sort_values(ascending=False) - 1)
            tmpdf = pd.DataFrame({
                "region_name": region_name,
                "sum_cosine_similarity": sum_cos_sim
            }, index=[0])
            cosdf = pd.concat([cosdf, tmpdf], ignore_index=True)

        cosdf.sort_values(by="sum_cosine_similarity", ascending=False, inplace=True)
        # too_regions = cosdf[cosdf["sum_cosine_similarity"] <= 0].region_name.tolist()
        too_regions = cosdf[cosdf["sum_cosine_similarity"] == 0].region_name.tolist()

        sum_preddf = pd.DataFrame(
            data = ['Breast', 'CRC', 'Gastric', 'Liver', 'Lung', 'WBC', 'no data'],
            columns = ['prediction']
        )

        all_preds = [
            item for item in pathlib.Path(path_to_03_output).glob(f"*{sampleid}/microAtlas_data/region_*/*_prediction_on_*.xlsx")
        ]

        predictiondf = pd.DataFrame()
        for file in tqdm(all_preds):
            tmpdf = pd.read_excel(file, index_col = [0])
            predictiondf = pd.concat([predictiondf, tmpdf], axis = 0)
        predictiondf = predictiondf[predictiondf["region"].isin(too_regions)]
        predictiondf = predictiondf.reset_index(drop = True)
        tmp_sum_preddf = predictiondf.groupby("prediction")["loocv_val_sample"].count().reset_index()
        tmp_sum_preddf.columns = ["prediction", sampleid]   
        sum_preddf = sum_preddf.merge(tmp_sum_preddf, right_on = "prediction", left_on = "prediction", how = "left")
        
        all_predictiondf = all_predictiondf.merge(sum_preddf, right_on = "prediction", left_on = "prediction", how = "left")
        all_predictiondf.fillna(0, inplace=True)
        all_predictiondf.to_excel(
            os.path.join(path_to_03a_output, f"LOOCV_res.xlsx")
        )
else:
    all_predictiondf = pd.read_excel(
        os.path.join(path_to_03a_output, f"LOOCV_res.xlsx"),
        index_col=0
    )    


all_predictiondf = all_predictiondf[all_predictiondf["prediction"].isin(["no data", "WBC"]) == False]

all_predictiondf.set_index("prediction", inplace=True)
all_predictiondf = all_predictiondf.T

# Find the column(s) with maximum value for each row
all_predictiondf['prediction_result'] = all_predictiondf.apply(
    lambda row: ', '.join(row[row == row.max()].index.tolist()), 
    axis=1
)

all_predictiondf = all_predictiondf.reset_index().rename(columns={"index": "LABCODE"})

all_predictiondf = all_predictiondf.merge(metadata[["LABCODE", "TYPE"]], 
                                          right_on = "LABCODE", left_on = "LABCODE", how = "left")

all_predictiondf[all_predictiondf["TYPE"] != "Control"]
